Add template for period choosing
Whenever the user selects a class (no different periods should show)
a new dialog pops up that allows the user to register to a specific
period for that class instead of having to scrooll though a huge list
to fid their class and period. When the user clicks the Create class
option in this new dialog, the creaet class activity needs to launch
with the class name filled so as to avoid errors.